site: Add competition overview
Add an overview of the competition. This is just the league standings (for
all divisions) and recent matches. For some leagues, the divisions can take
up a lot of space, but some leagues only have e.g.
8 teams in 2 divisions, so I don't know if there's a general solution. h3s
and h4s look a little odd with the default spacing, so reduce it a bit.

Signed-off-by: Sean Anderson <seanga2@gmail.com>