importer: link_matches: Don't consider maps when matching
Firstly, the comparision should have been "... OR mapids = '{}'" which
caused matches without maps not to match correctly. Secondly, the maps are
often incorrect. This can be because of combined logs, but sometimes the
map is ALL CAPS. Combined logs aren't a problem (since we will ignore
them), but we're really not gaining anything by matching on maps.

Signed-off-by: Sean Anderson <seanga2@gmail.com>