site: team: Add matches page
Add a page for the matches a team has played in. I would have liked to have
reused the SQL from common (and maybe I will in the future), but the
dynamic SQL required would have been too much IMO.  We can reuse the
match_table macro though, so do so.

Signed-off-by: Sean Anderson <seanga2@gmail.com>