site: comp: Add class to players page
Add the primary class to the players page. Technically, the list of all
classes is right there to be aggregated. However, it would be such a pain
to integrate without messing up the other stats without joining to
class_stats. I suspect that would kill query performance, so I have gone
with the primary class.

Signed-off-by: Sean Anderson <seanga2@gmail.com>