Fix negative durations
The aggregate statistics we employ do not handle negative durations
partiocularly well. This is most visible on the leaderboard, where several
players are highly ranked with negative total playtime. Determine the
correct playtime as best we can, and clamp to zero where we can't. This
isn't great, but it's better than nothing.

Closes: #94
Signed-off-by: Sean Anderson <seanga2@gmail.com>