migrations: Update demo migration to match link_demos
Back in commit 0533877 ("importer: Don't link demos without players"), I
modified the condition for linking demos. Update the SQL migration as well,
since I still run it occasionally (such as after importing many new logs or
demos). Also update the migration to create a temporary table with the
demos to link, as this is much more efficient in postgres.

Signed-off-by: Sean Anderson <seanga2@gmail.com>