site: team: Show team name and avatar from most-recent competition
Instead of showing the name/avatar from the most-recently-update
competition, show them from the most-recent competition overall. However,
we still need to get the max of fetched, so sort in array_agg instead of a
global order by.

We also can skip getting the compid, since it was only used for teams with
league_team_per_comp, but we want to show all league_teams on the team
page.

Signed-off-by: Sean Anderson <seanga2@gmail.com>