Add RGL support
At long last, add RGL support. There are still a lot of API bugs and data
errors on the RGL side, requiring several hacks on our end:

- There is no way to differentiate rounds except by name. We must create
  our own round_seq upon insertion. There are still several cases where
  rounds have the same name...
- Most teams have incorrect divisions, so we must fix them up from the
  match data. This doesn't always work, so there are several special cases.
- Sometimes the timestamp for when a player joined is earlier than for when
  they left. We just ignore these for now.

The rest of this commit is fairly straightforward. The table checks are
adjusted a bit to reflect the data as-is. Additionally, some of the indices
need to be modified to still apply.

The site itself is mostly the same. The team page is adjusted to link to an
appropriate rgl_teamid (or the most-recent one).

At the moment there is no scheduled import, since the round stuff is still
unresolved. I hope that this can be addressed quickly so I can deploy this
feature.

Signed-off-by: Sean Anderson <seanga2@gmail.com>