site: league: team: Add players page
Add a players tag to the team page. This makes it easy to compare players
in a team. Unlike the comp version, we only show stats for players on the
team. IMO this makes most sense, since we don't usually want to compare
e.g. the best enemy player in our match.

Closes: #96
Signed-off-by: Sean Anderson <seanga2@gmail.com>