site: teams: Fix incorrect ordering of matches on team overview
There can only be one call to get_order per page. This is because we cache the selected order for
the sort macro. Therefore, we can't call get_order from helpers (such as get_comps and get_matches)
which might be called multiple times (such as on the overview page). Provide an explicit default
order clause instead.

Fixes: 513352c ("site: league: team: Add competitions tab")
Signed-off-by: Sean Anderson <seanga2@gmail.com>