site: player: Show first as well as last competition under teams
It's a bit confusing to show only the most-recent competition for a player,
especially if they have been rostered for multiple competitions. Show the
last-recent competition as well, to give a better overview of how long they
have been on the team.

Finding the maximal/minimal row is such a pain to do in SQL. Switch from
the previous method of using window functions to using a CTE and multiple
selects. While window functions work fine for grabbing either the minimum
or maximum, they are really terrible for grabbing both (especially when
there are NULL rows to ignore. To ensure we don't do a sequential scan on
matches, denormalize the first/last scheduled date into competition.

Signed-off-by: Sean Anderson <seanga2@gmail.com>