importer: Don't crash when there's an error getting a match
When there's a network (or other) error getting a match, we return None.
This is fine for methods which yield results, since that turns into an
empty iterator. But for regular methods that means we have to check for
None every time. Just re-raise the exception and catch it later on.

Fixes: ad2f5d7 ("Add RGL support")
Fixes: 481a320 ("Support importing ETF2L matches")
Signed-off-by: Sean Anderson <seanga2@gmail.com>