Add updated field to logs
Add an updated field to logs, which is updated when the log is modified.
This generally occurs when we link a demo or match to the log.

Support using this field whenever we call last_modified. There are a few
places (players, teams) where we show log data on the page but this doesn't
factor into the last_modified field. This is because we would need to join
on the logs to determine this, which is expensive. I am deferring this
until I determine how to implement this efficiently.

Now that we have it, add the updated field to the logs API, and add an
updated_since filter as well. This should help clients query us
efficiently.

Signed-off-by: Sean Anderson <seanga2@gmail.com>