Switch to pg_basebackup
pg_dump has a heavy performance cost on the database. Switch to
pg_basebackup which is much lighter.  Whereas pg_dump produces a logical
dump of the data in the database, pg_basebackup produces a tar with all the
files in the tablespace. As such, the restoration process is just
extracting the files to an appropriate location.

Closes: #105
Signed-off-by: Sean Anderson <seanga2@gmail.com>