site: Add medic leaderboard
In the same style as the generic leaderboard, add a leaderboard for medic
stats. The stats need a bit of massaging, especially since so many are
optional or derived. The info is not the best since it's so easy to mess up
the stats by e.g. running vaccinator. But it's better than drops.tf :)

Signed-off-by: Sean Anderson <seanga2@gmail.com>