site: Add minimum logs filter to leaderboards
Often it can be difficult to filter out uninteresting data when sorting
by columns which are not "naturally resistant" to outliers. For example,
it is quite difficult to find players with a high drop percentage
without excluding the numerous players who dropped every built uber in
the only game they ever played. To address this, add a filter for
minimum logs.

Signed-off-by: Sean Anderson <seanga2@gmail.com>