site: Vendor javascript dependencies
Download our javascript dependencies when building the site. This allows
us to improve load times since the browser can reuse the same
connection. It also mildly improves privacy.

I also looked at webpack, but it seemed to be overkill for what I want
to do. Most of the files on this site are already split into independent
bundles included only when necessary. And it also pull in a lot of
javascript depenencies/overhead. I think it would still be nice to run a
tree shaker on our dependencies.

Closes: #59
Signed-off-by: Sean Anderson <seanga2@gmail.com>