importer: rgl:
The RGL API has no way to specify that we only want matches updated
since the last time we fetched something, so currently we fetch logs
based on the most-recently-scheduled match (that we know about). This
works since only completed matches are imported. However, the RGL admins
in their infinite wisdom decided to forfeit some matches scheduled for a
month in the future, breaking updates. Fix this by preventing the
most-recently-scheduled match from being newer than the current time.

Signed-off-by: Sean Anderson <seanga2@gmail.com>