cache: Use a sentinel for objects about to be cached
With the existing protocol, stale objects could remain in the cache if they
were being computed while an update was in progress:

importer            wsgi
=================== ===================
                    cache.get()
                    cur.execute(SELECT)
cur.execute(UPDATE)
cache.delete()
                    cache.put()

prevent this by inserting a sentinel object before filling the cache. If
the sentinel is absent after we've generated the object, then the database
has been updated in the meantime. Therefore we only update the cache if the
sentinel remains unchanged.

While we're at it, refactor the caching logic into a context manager to
make it easier to reuse, and move the tests to their own file.

Fixes: c48c718 ("importer: Purge player overview cache when updating players")
Signed-off-by: Sean Anderson <seanga2@gmail.com>