site: log: Cache logs and merge the results in python
AI scrapers really love the log endpoint since there are a bunch of ways to
add/remove logs. To improve performance, add some caching. This is done
per-log so that removing a log (or adding another) will mostly hit the
cache. Unfortunately, this means we have to do the aggregations in python.
While far more verbose than SQL, this ends up only being around 150 lines
more, mostly because the SQL can also be simplified.

To aid in development, the data passed to the template is exactly the same
as before this commit. The next commit will refactor things to simplify the
template (now that we have some more flexibility on our data structure).

Signed-off-by: Sean Anderson <seanga2@gmail.com>