site: Cache player logs
Cache players' logs. The query parameters are encoded into the name as a hash. To invalidate old
keys, the primary overview cache entry has a random "version." This is used as a key for the hash,
ensuring that we use different keys when the overview is purged.

32 bytes of data is probably way overkill for the version. We use base64
instead of base16 to keep the key size down. I considered base85. but it
was around twice as slow. Blake2b is used for the hash since it has a
convenient key parameter, and has a configurable digest size (so we can
get an even number of characters).

Signed-off-by: Sean Anderson <seanga2@gmail.com>