site: Reduce query timeout
The performance of the site has improved and queries no longer take nearly
as long as they used to.  The maximum I observed in sentry was 15 seconds,
with a 95th percentile of 5 seconds. Reduce the per-query timeout to 20
seconds, and reduce the dummy cache timeout to just 30. With the default
settings, values that expire in under 60 seconds will never be removed by
memcached's LRU crawler.

Signed-off-by: Sean Anderson <seanga2@gmail.com>