site: Set ETag for player requests
Set the etag based on the version of the player in the cache, which serves
as a stronger indicator of cache validity than last-modified. The latter
may not always be updated on changes, but the etag will change when the
cache is purged.

128 bits of data are enough to prevent collisions in non-adversarial
conditions, so reduce the length of the version (and the ETag) to 16
bytes.

Signed-off-by: Sean Anderson <seanga2@gmail.com>