site: Cache index page
Cache the index page and invalidate it whenever we update a log or a
player. Before this commit, only cache for players' pages would need to be
invalidated, but players imported via RGL or demos would return 404 because
they had no logs. Therefore, add some more purges to demos and RGL.

Signed-off-by: Sean Anderson <seanga2@gmail.com>