site: log: Fix rendering matches where all logs are duplicates
There are around 10 matches where all of the linked logs are duplicates.
From what I can tell this usually happens when there is something that
causes the earlier log(s) not to match (such as rounds from the warm-up
being included). Then someone uploads an edited log that matches with the
match.

I'm nopt really sure what to do with these in general, as there are not
very many of them. For now, just ensure we don't crash if full_logs is
None.

Fixes: 667bbab ("site: log: Only get each match once")
Signed-off-by: Sean Anderson <seanga2@gmail.com>