site: player: Don't cache logs
This cache has a very low hit rate, so disable it for now. I think a better
approach could be to cache e.g. 1000 logs when someone requests page 2 so
that pages 3-10 can be served from cache. But that needs testing.

This mostly reverts commit 5e79e3d ("site: Cache player logs"), although
the hast_object stuff is left in in case I want to

Signed-off-by: Sean Anderson <seanga2@gmail.com>