site: Set cache-control
Control nginx's behavior through the cache-control header. This lets us
customize max-age based on the last modification time. This is good enough
for several pages Most pages that don't set last-modified likely update
around once a day ish so set the max age to 5 minutes. The exception is
root.logs, which updates every time a lot is uploaded. root.log is a
favorite of bots and it is handeled much better my memcached, so don't
cache that one in nginx.

Signed-off-by: Sean Anderson <seanga2@gmail.com>